Tighten MoDT.optimizeC() condition number check
MoDT.optimizeC() is responsible for making sure that the fitted C matrix
satisfies the max_cond condition number threshold. However, due to numerical
innacuracies (and/or the difference between 1- and 2-norm condition
numbers), this was able to produce C matrices that didn't satisfy the
setParams() condition number check. This could lead to a situation where we
fit a MoDT model without problems but are unable to copy these values into a
new model using setParams().

To mitigate this problem, we tighten the condition number check inside
optimizeC (by 5%) in order to compensate for the potential difference.