Cast weight vector to same datatype as spikes
Previously, attachData() did not check that the weight vector `w` was the same
datatype as the given spikes `Y`, which could cause an error when performing
model operations later. This was particularly annoying when accepting the
default value of `w`. This commit makes attachData() automatically cast `w` to
the same datatype as `Y`.