Improve numerics by using sum(log) vs log(product)
Mathematically, the determinant of X is equivalent to the product of the
diagonal terms of the Cholesky decomposition of X. In other words,
    log(det(X)) = log(prod(diag(chol(X)))) = sum(log(diag(chol(X)))).
Previously, I chose to use log(det(X)) because it was faster, but now that
we're using single-precision data, we are getting cases where
    det(X) > realmax('single') = 3.4e38,
in which case MATLAB returns det(X) = Inf. This commit addresses this issue
by using the more numerically-stable sum(log(diag(chol(X)))) instead.