Bugfix (ensure return datatype matches arguments)
This commit fixes a bug where calling this function with single-
precision arguments would produce double-precision results.

This bug is unlikely to have caused any issues, however, because it is
essentially unreachable. This code is only executed when use_mex=true
but the corresponding MEX file is unavailable, but it is difficult to
enter that state because you can't set use_mex=true unless the MEX
files are present.